#3 Draft question server storage
Draft questions are now stored in the database

+++ KiwiBoard/app/src/main/java/com/example/kiwiboard/ProfReportFragment.java
@@ -34,7 +34,7 @@ public class ProfReportFragment extends Fragment {
         ArrayList<Question> questions = currentcourse.getQuestions();
         Question question;
         double pointsreceived, percentage;
-        int maxpoints;
+        double maxpoints;
         ArrayList<Double>averages=currentcourse.calculateAverages();
 
         // Class Average
